requirements.txt
python-binance==1.0.29
python-dotenv==1.1.1
pandas>=1.5.0
numpy>=1.21.0
ta==0.11.0

